refactor(gui-v2): move useGlobal types to useGlobal dir
