+++ packages/nc-gui-v2/composables/useGlobal/getters.ts
@@ -1,8 +1,7 @@
-import type { UseGlobalStateReturn } from './state'
-import type { Getters } from '~/lib'
+import type { Getters, State } from './types'
 import { computed } from '#imports'
 
-export function useGlobalGetters(state: UseGlobalStateReturn) {
+export function useGlobalGetters(state: State) {
   /** Getters */
   /** Verify that a user is signed in by checking if token exists and is not expired */
   const signedIn: Getters['signedIn'] = computed(
@@ -10,9 +9,9 @@ export function useGlobalGetters(state: UseGlobalStateReturn) {
       !!(
         !!state.token &&
         state.token.value !== '' &&
-        state.payload.value &&
-        state.payload.value.exp &&
-        state.payload.value.exp > state.timestamp.value / 1000
+        state.jwtPayload.value &&
+        state.jwtPayload.value.exp &&
+        state.jwtPayload.value.exp > state.timestamp.value / 1000
       ),
   )
